Inline datashader
